+++ src/00_dev/index.html
@@ -24,15 +24,10 @@
         <div class="macbook-pro-16-1">
           <div
             id="div1"
-            data-webgl="10_s"
+            data-webgl="10_slide01"
             data-tex-1="/img/output9.jpg"
             data-tex-2="/img/output0.jpg"
           ></div>
-          <!-- <img
-              class="samuel-ferrara-1527pjeb6jg-uns-icon"
-              alt=""
-              src="../img/output9.jpg"
-            /> -->
         </div>
       </div>
     </div>
