--- src/00_dev/scripts/glsl/10_s/index.js
@@ -1,18 +1,0 @@
-import gsap from "gsap";
-
-import vertexShader from "./vertex.glsl";
-import fragmentShader from "./fragment.glsl";
-
-import { CustomObject } from "../CustomObject";
-
-class ExtendObject extends CustomObject {
-  fixVertex() {
-    return vertexShader;
-  }
-
-  fixFragment() {
-    return fragmentShader;
-  }
-}
-
-export default ExtendObject;
