+++ src/00_dev/styles/style.scss
@@ -1,12 +1,3 @@
-// .samuel-ferrara-1527pjeb6jg-uns-icon {
-//   align-self: stretch;
-//   flex: 1;
-//   position: relative;
-//   max-width: 99%;
-//   overflow: hidden;
-//   max-height: 99%;
-//   object-fit: cover;
-// }
 #div1 {
   align-self: stretch;
   flex: 1;
@@ -28,11 +19,5 @@ img {
   height: 100vh;
   overflow: hidden;
   display: flex;
-  // flex-direction: column;
-  // align-items: center;
-  // justify-content: center;
-  // min-width: 700px;
-  // max-width: 1800px;
-  // min-height: 700px;
-  // max-height: 1200px;
+  justify-content: center;
 }
